Apply patch by George - gnassas (at) mac.com. Provides a cleaner way to make non blocking connections to the BE. Updates mythwelcome to use the new method to create its non blocking connection.
Changes by me to update mythlcdserver to use the same method to
connect to the BE.

Updates the protocol to version 22.

These changes are basically the start to get mythwelcome working on remote
frontends as well as combined FE/BE machines. There are still a few issues
to fix though.

Closes #617

git-svn-id: http://svn.mythtv.org/trunk@7883 4b91aca1-2a1f-0410-a73a-a78f2df169e8