Fixes #675.
This adds a parameter, called 'local_only', to the PreviewGenerator
constructor which defaults to true. When set to true the 
preview generator will never open a connection to the backend,
even if it can't find the recording to preview locally.

Also added some doxygen documentation that tells developers using
the PreviewGenerator that this should always be set when using 
the preview generator on the backend.



git-svn-id: http://svn.mythtv.org/trunk@7904 4b91aca1-2a1f-0410-a73a-a78f2df169e8