inuseprograms related stuff.
Convert inuseprograms table to store a hostname and an identifier saying
how the recording is being used on the host (ie, 'recorder', 'player',
'flagger', etc.)

Modify PlaybackBox to allow deleting of recordings that are not being
played and not recording LiveTV.

Modify all places NuppelVideoPlayer is used to pass in a proper identifier
for the inuseprograms table so PlaybackBox can know what recordings it can
delete.

Re-enable updating of inuseprograms table for mythcommflag and mythtranscode.
These will not block deletion of a recording, so it may come in handy to
know if one of these is using a recording.



git-svn-id: http://svn.mythtv.org/trunk@7928 4b91aca1-2a1f-0410-a73a-a78f2df169e8