Increases buffering to 1.5 seconds if kernel memory is available.
The increase a few days ago was to 550 milliseconds, but it wasn't
very smart about the bitrate of different stream types, this assumes
a lower bitrate for non-video streams and so allocates less kernel
memory overall, but allocates 3x more for the video stream. If it 
can't allocate the needed kernel memory we try smaller sizes until
it works and then print out a warning with the approximate number
of milliseconds of buffering we do have.



git-svn-id: http://svn.mythtv.org/trunk@7945 4b91aca1-2a1f-0410-a73a-a78f2df169e8