References #551.
This makes the eitscanner a little smarter about handling multiple
EIT scanning cards. It adds randomness to when scanning starts and
when channel changes happen, this should make it much less likely
that two or more cards act in lockstep when it comes to changing
channels and calling for reschedules.

This also implements rate limiting across all eitscanners on the 
same backend on the calls to reschedule.

This also adds randomness to which channel the scanners start at.
This just assures that when multiple scanners first start up two
or more of them aren't scanning the same exact channels. This doesn't
address Brandon's problem, but only makes the scan faster right after
you first start the backend if you do have multiple cards. After
a while the scanners would be out of lockstep anyway, this just
speeds the process.



git-svn-id: http://svn.mythtv.org/trunk@7946 4b91aca1-2a1f-0410-a73a-a78f2df169e8