Added the concept of playback groups (aka playgroups).  Like recording groups (aka recgroups), each recording is put into a playgroup when it is recorded as specified in its recording rule.  When that recording is watched, selected playback options are set based on the playgroup instead of using system or host specific settings.  For example, you can set the initial timestretch speed to 1.5x for "Talking Head" news programs or set the skip ahead time to 25 seconds for football.
Initially, the only settings changed to use playgroups are skip ahead
and skip back times and initial timestretch speed.  Others will
probably be changed in the future.

Besides the normal values, each playgroup setting can be set to the
special "(default)" value.  When a playgroup setting is requested,
Myth will first look in the specified playgroup, then in the "Default"
playgroup and finally a hard-coded value.  The first one not set to
"(default)" is used.

When a new recording rule is created, Myth will automatically set the
playgroup for the rule if the program title or category exactly
matches the name of a playgroup.  Additionally, each playgroup has a
titlematch regular expression.  Myth will also automatically set the
playgroup if the program title matches this regular expression.


git-svn-id: http://svn.mythtv.org/trunk@7965 4b91aca1-2a1f-0410-a73a-a78f2df169e8