Cosmetic debugging changes (still only visible with "-v important" though)
In dbcheck, print the current DB schema version whether we need to upgrade
or not.

In TVRec::GetStartChannel, make error messages a little more clear.  There
is no reason to say "Start channel '' invalid" when the startchan is always
empty at that line of code, so get rid of the ''.



git-svn-id: http://svn.mythtv.org/trunk@8012 4b91aca1-2a1f-0410-a73a-a78f2df169e8