Closes #698.
We really should be looking at the Visual rather than the depth for several things in videoout_xv, including this. But since this is just a sanity check for a not-yet-fully-implemented feature, I've relaxed the requirement a bit.

Please put future chromakey bugs reports on the "Chromakey OSD" task ticket #512 until that ticket is closed.



git-svn-id: http://svn.mythtv.org/trunk@8022 4b91aca1-2a1f-0410-a73a-a78f2df169e8