References #667.
 * Adjust for aspect ratio effect on font scale.

I believe this is due to the fonts being scaled on hmult alone in
OSD instead of the scale in x&y. This adjustment appears to work
for video aspect ratios in the range 1.3333 - 2.0. The font size 
is now proportional to the vertical size of the visible video 
window, which appears to work well with my test streams.



git-svn-id: http://svn.mythtv.org/trunk@8030 4b91aca1-2a1f-0410-a73a-a78f2df169e8