Fixes #616.
This should make DVB drivers that report the signal improperly
work better, and still be compatible with well implemented drivers.

The problem is really in the DVB API v 3.x being unclear for
both the signal and signal-to-noise ratio. This isn't properly
addressed in the 4.0 API, but the API has been changed to better
reflect practice so I'm using it for interpreting these values.
The proposed 4.0 API still does not specify how the drivers 
should calculate these values; but any one looking at both 
the 3.1 and 4.0 API's should see that these values should be
in the range 0..32767 so that they will always change 
monotonically with respect to signal quality changes for 
anyone implementing either API.



git-svn-id: http://svn.mythtv.org/trunk@8038 4b91aca1-2a1f-0410-a73a-a78f2df169e8