Fix missing "virtual" on declaration of VideoOutput::ClearAfterSeek().
During the investigation of #710, I found that VideoOutputXv::ClearAfterSeek()
was never getting called when NVP calls videoOutput->ClearAfterSeek(). Instead,
it appears that VideoOutput::ClearAfterSeek() was always being called since
it was not declared virtual.  I don't know if this cures any outstanding
XvMC issues, but it might since the XvMC code at the bottom of
VideoOutputXv::ClearAfterSeek() was never being executed.



git-svn-id: http://svn.mythtv.org/trunk@8062 4b91aca1-2a1f-0410-a73a-a78f2df169e8