Makes previewgenerator smarter about finding recordings locally.
If the pathname is not local, we regenerate it from the basename
plus the prefix and check if it exists locally. If the generated
variant exists locally we use that instead. This should address
one of the problems Petr Stehlik ran into, where a preview was
not generated at the end of a recording and the preview generator
complained about not finding the file.



git-svn-id: http://svn.mythtv.org/trunk@8088 4b91aca1-2a1f-0410-a73a-a78f2df169e8