This makes prebuffering pauses a little easier to debug.
This prints out the status of each frame when a prebuffering
pause happens. If many the frames are 'A'vailable it means
that the decoder thread is starving the display thread, if
most of frames are in any of the other states ('d'isplaying,
'U'sed (prebuffered), 'L'imbo, or 'p'ause) then it is the
display thread that is starving the decoder thread. The
latter usually only happens with XvMC. If say ten frames
were 'd'isplaying you would know that there was a leak
in the display code. The same is true with the XvMC frame
states of 'p'ause, and 'L'imbo.

There is also an additional error message when the decoder
thread dies, and a couple additional decoder thread status
messages when '-v playback' is used.



git-svn-id: http://svn.mythtv.org/trunk@8117 4b91aca1-2a1f-0410-a73a-a78f2df169e8