Disables MPEG-TS Reset.
I don't think this is needed anymore. I believe this was used to
re-read the header in the mpeg-ts file after we started recording
a new mpeg file in the same ringbuffer in the old LiveTV 
implementation. Plus it is causing problems on quick channel
changes with the new LiveTV implementation.



git-svn-id: http://svn.mythtv.org/trunk@8145 4b91aca1-2a1f-0410-a73a-a78f2df169e8