Fixes #767.
The problem was that the DVBRecorder would completely ignore audio
streams if none was found in a preferred language. This fixes this
by re-running the stream selection without language preferrences if
no audio was found the first time around. Actually, I backported
the AutoPID function from the #712 patch, so the diff is a little
bigger than it would be otherwise.



git-svn-id: http://svn.mythtv.org/trunk@8188 4b91aca1-2a1f-0410-a73a-a78f2df169e8