MythTV PlaybackBox fixes.
Remove a recording from the Playlist if we delete the recording.

Make the DELETE keybinding only try to delete the currently highlighted
recording.  Previously it would try to delete the playlist and fail
miserably, causing mythfrontend to appear to lockup because of focus
issues.  If you want to delete all items in a playlist, use the delete
button under the playlist menu.

Don't ask the user if they want to "delete it, but allot it to re-record"
when deleting a recording from the LiveTV recording group.  These recordings
are not tracked in oldrecorded anyway, so there's no need to ask.  This was
Oscar's catch, I'm just committing it after talking with him.



git-svn-id: http://svn.mythtv.org/trunk@8197 4b91aca1-2a1f-0410-a73a-a78f2df169e8