References #712.
This is part one of two sets of innocuous changes to the DVBRecorder pre-DeviceReadBuffer patch. The idea here is to seperate things that are unlikely to break things from the things that may, so there are no new features or bug fixes in this commit.

This adds a IsOpen() call, it separates writing regular packets and PAT/PMT packets, it renames the and rearranges some methods and variables, etc.



git-svn-id: http://svn.mythtv.org/trunk@8203 4b91aca1-2a1f-0410-a73a-a78f2df169e8