References #712.
This commit could cause problems. It changes how PIDs are tacked,
instead of using 5 QMaps and one vector to keep track of information
about the streams, this creates one entry in one QMap to track the
information. This means 4 less QMap lookups per TS Packet processed
by the recorder.

This also changes the TS->PS transform to use the same keyframe tracking
as the TS recordings, and to use the same source for constants instead
of using it's own defines.



git-svn-id: http://svn.mythtv.org/trunk@8206 4b91aca1-2a1f-0410-a73a-a78f2df169e8