Fixes some problems seen when looking at #799.
We were parsing the sequence header incorrectly when looking
at the height. Also we might as well use the reported fps
if we get a sequence header. And finally we should really
check if we have enough buffered data before we go reading
data past bufptr.

There are also some sanity checks, like if ffmpeg disagrees
with the sequence header on the framerate a warning is printed
out when using '-v playback'. I also made the HandleGopStart
code a little more self-documenting.

I've removed the sanity check (width < 2500 && height < 2000),
since this was probably only needed due to the broken sequence
header parsing.



git-svn-id: http://svn.mythtv.org/trunk@8259 4b91aca1-2a1f-0410-a73a-a78f2df169e8