References #802.
2nd patch from nooneimprt4nt AT gmail , com

This fixes the problem, so long as the channum's are unique across
all inputs on a card.

nooneimprt4nt, can you look at SwitchToInput() and see if you could
make it perform the DiSEqC switch to the specified input even if
a channum exists on two inputs? You probably need to make some
modifications to GetChannelOptions()...



git-svn-id: http://svn.mythtv.org/trunk@8277 4b91aca1-2a1f-0410-a73a-a78f2df169e8