mythfrontend statusbox changes and related remoteutil.cpp cleanup...
Add a new QUERY_GETEXPIRING backend query to get the list of all programs
which may be expired and the order they will be expired in.

Consolidate some duplicate code in remoteutil.cpp from all the functions that
request a list of programs from the backend.

Add a new section on the Status page in Mythfrontend to display the list
of programs that can be expired and show how much disk space they are using
up.  The programs are listed in the order they will be expired.

fix a bug in the Status page that would not let the user scroll to view
detail on the machine status page if the content box on the right hand side
was not full.

On the Status page, allow the MENU key to bring up the popup menu for
the jobs in the queue, previously only SELECT would popup the menu.



git-svn-id: http://svn.mythtv.org/trunk@8280 4b91aca1-2a1f-0410-a73a-a78f2df169e8